Moving certain files to be in line with how the project is structured and adding a small development boolean for changes page
