+++ lib/PageSelectorTypes.d.ts
