+++ pages/blog/[id].tsx
@@ -7,7 +7,7 @@ import { getMDXComponent } from "mdx-bundler/client";
 import { useMemo } from "react";
 
 import { GetStaticProps, GetStaticPaths, PreviewData, InferGetStaticPropsType } from "next";
-import { PostParams, PostStaticPropsResultProps, PostFrontMatter } from "../../lib/id";
+import { PostParams, PostStaticPropsResultProps, PostFrontMatter } from "../../lib/idTypes";
 
 const development = process.env.NODE_ENV === "development";
